Fixes missing line in sidekiq worker, changes db to Postgres to use money data type, adds csv import and export functionality

+++ db/migrate/20180926194206_create_rules.rb
@@ -2,7 +2,7 @@ class CreateRules < ActiveRecord::Migration[5.2]
   def change
     create_table :rules do |t|
       t.string :relative_price
-      t.float  :discount
+      t.money  :discount
       t.string :reporting_make
       t.string :brand_name
       t.string :category_name
