+++ db/migrate/20180927035951_create_discounted_parts.rb
@@ -1,7 +1,7 @@
 class CreateDiscountedParts < ActiveRecord::Migration[5.2]
   def change
     create_table :discounted_parts do |t|
-      t.float :price
+      t.money :price
       t.string :sku_label
       t.string :reporting_make
       t.string :brand_name
